Diagnostic list cleanup
